+++ roles/assetmgr.rb
@@ -1,8 +1,0 @@
-name 'assetmgr'
-description 'AssetMgr Rails Application'
-run_list(
-  'role[basic]',
-  'role[nginx]',
-  'role[postgresql]',
-  'recipe[assetmgr]'
-)