+++ roles/basic.rb
@@ -1,5 +1,6 @@
 name 'basic'
 description 'Basic configuration'
 run_list(
-  'recipe[apt]'
+  'recipe[apt]',
+  'recipe[git]'
 )