+++ roles/testapp.rb
@@ -1,0 +1,7 @@
+name 'testapp'
+description 'Test Rails Application'
+run_list(
+  'role[basic]',
+  'role[postgresql]',
+  'recipe[testapp]'
+)